refactor(storage): share RocksDB adapter with filesystem
