+++ .changenotes/share-filesystem-rocksdb-storage.md
@@ -1,0 +1,6 @@
+---
+type: patch
+---
+
+Filesystem-backed Lix instances now use the shared RocksDB storage adapter so
+RocksDB performance improvements apply consistently across storage modes.
