+++ packages/local-filesystem/Cargo.toml
@@ -19,16 +19,7 @@ workspace = true
 
 [features]
 default = []
-rocksdb = ["dep:bytes", "dep:rocksdb", "dep:tempfile", "dep:tokio"]
+rocksdb = ["dep:lix_rocksdb_storage"]
 
 [dependencies]
-lix_engine = { workspace = true }
-
-bytes = { version = "1", optional = true }
-rocksdb = { version = "0.24", default-features = false, features = ["bindgen-runtime"], optional = true }
-tempfile = { version = "3", optional = true }
-tokio = { version = "1", features = ["rt", "sync"], optional = true }
-
-[dev-dependencies]
-bytes = "1"
-tempfile = "3"
+lix_rocksdb_storage = { workspace = true, optional = true }
