+++ packages/local-filesystem/src/lib.rs
@@ -1,10 +1,10 @@
-//! Filesystem-specialized persistence for Lix.
+//! Compatibility exports for filesystem persistence.
 //!
-//! This internal crate owns the RocksDB implementation behind
-//! `lix_sdk::LocalFilesystem`.
+//! `lix_sdk::LocalFilesystem` uses the shared RocksDB storage adapter so all
+//! RocksDB behavior and optimizations have one implementation.
 
 #[cfg(feature = "rocksdb")]
-mod rocksdb;
-
-#[cfg(feature = "rocksdb")]
-pub use rocksdb::{RocksDBFilesystem, RocksDBFilesystemRead, RocksDBFilesystemWrite};
+pub use lix_rocksdb_storage::{
+    RocksDB as RocksDBFilesystem, RocksDBRead as RocksDBFilesystemRead,
+    RocksDBWrite as RocksDBFilesystemWrite,
+};
