+++ packages/rocksdb-storage/Cargo.toml
@@ -22,3 +22,7 @@ lix_engine = { workspace = true }
 bytes = "1"
 rocksdb = { version = "0.24", default-features = false, features = ["bindgen-runtime"] }
 tempfile = "3"
+tokio = { version = "1", features = ["sync"] }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt"] }
